[applets/batterymonitor] Put the toolbar in the header area
Summary:
This patch implements part of the mockup at T10470 by putting the power management
checkbox and settings button in a heading area visually connected to the titlebar.

In the process, I fixed a problem with the power management checkbox to make the layout's
spacing work: the checkbox now has text of its own, instead of living in a mouse area
with a separate label.

Test Plan:
{F8254807}
{F8254808}

Reviewers: #vdg, #plasma, broulik, niccolove, manueljlin

Reviewed By: #vdg, manueljlin

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T10470

Differential Revision: https://phabricator.kde.org/D29116