BoorkmarksRunner: Set QIcon instead of Favicon pointer in BookmarkMatch
Summary:
Before: A pointer to the favicon gets written in the BookmarkMatch and using this pointer the icon gets created when the Plasma::QueryMatch gets created.
Now: Using the favicon pointer the icon gets created and written to the BookmarkMatch, when the QueryMatch is created the icon value is read.

This makes the code easier to understand and the BookmarkMatch is independent of the Favicon class.

Test Plan: Should compile and display icons.

Reviewers: apol, broulik, davidedmundson

Reviewed By: davidedmundson

Subscribers: bruns, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28439