BookmarksRunner: Early return and formatting
Summary:
As requested in D28196 this patch handles the early return and the formatting.
Additionally QRegularExpression has been used instead of QRegExp and QStringLiteral instead of QString.

Test Plan:
Should compile and work exactly as before.

To test migration to QRegularExpression:
create multiple profiles
delete the `dbfile` entry from the general group in .config/kdeglobals
restart krunner and make sure that the value of the dbfile entry is the path to the default profile.
The bookmarks should also be displayed.

Reviewers: bruns, broulik, #plasma

Reviewed By: bruns

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28473