[System Tray] Always draw the highlight behind the icon
Summary:
There was a slight issue introduced with D28109: the systray highlight can now cover up the icons it's highlighting.

This isn't really an issue with Breeze because it uses a line for a highlight, but it is for themes that use a filled-style highlight.

I set a z value for the highlight that should ensure it's now drawn only behind the icon.

Test Plan:
Tested with the //Mondrian// Plasma theme

Before:
{F8280717}

After:
{F8280718}

Reviewers: kmaterka, #plasma

Reviewed By: kmaterka

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29380