[applets/systemtray] Show base icon when AttentionIcon not set
Summary:
If status is set to NeedsAttention it tries to use Attention Icon. If Attention Icon is not set, empty/null QIcon is set. QML can't check if QIcon is null or not, as a result it tries to use empty icon and nothing is rendered.
Set null QVariant if AttentionIcon is not valid so that QML check will work correctly.

Test Plan:
* Set NeedsAttention as a status and with no Attention Icon
* Base Icon should render correctly

Reviewers: #plasma_workspaces, #plasma, ngraham, broulik, davidedmundson, mart

Reviewed By: #plasma_workspaces, #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29544