Kill Runner: Remove duplicate function calls and refactor
Summary:
Duplicate function calls to load the config in the runner/config have been removed.
The actions of the runner are also refactored, similar to https://phabricator.kde.org/D3688.
Also the user (owner of process) has been removed from the subtext, because the second line is not shown and adding the user info to the same line would be too long.

Test Plan: Should compile and kill processes as before.

Reviewers: davidedmundson, ngraham, broulik, meven, apol

Reviewed By: meven

Subscribers: apol, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27513