[sddm-theme] Prevent the logo from leaking in after a fadeout
Summary:
Even when the vendor logo is turned off it can be briefly seen when the login screen changes from the fadeout state back into normal.

Comments in other code say that "OpacityAnimator when starting from 0 is buggy (it shows one frame with opacity 1)" so this patch replaces it with NumberAnimation.

There's also a minor fix included - placing the drop shadow sooner in the hierarchy because it should be drawn underneath an object, not in front of it.

Test Plan:
Reduce the fadeout time to something small and then:

`sddm-greeter --test-mode --theme /usr/share/sddm/themes/breeze`

Reviewers: ngraham, cblack, #plasma, mart

Reviewed By: ngraham, #plasma, mart

Subscribers: mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29351