[Free Space Notifer] Use critical notification instead of tray icon and monitor Root, too
This refactors the free space notifier module to use a critical (i.e. persistent and always on top) notification for warning of low disk space.
The, albeit blinking, tray icon is easy to miss, especially when you're running a full screen terminal, which I typically do while compiling stuff.
It now also monitors the Root folder, if it's on a separate partition from the user's home. Furthermore, Filelight is offered to explore the drive (if installed).

The overall warning logic remains pretty much the same:

* Once the drive goes below the configured threshold a warning notification is shown, it stays on screen until dismissed by the user or
  when free space is above warning threshold again.
* The notification is emitted again when free space drops below half the previous threshold, for added sense of urgency should the drive
  be rapidly filled up
* The notification is also emitted again if free space remains below the threshold for more than one hour

(It will only emit again when it was closed, obviously, so you won't end up with a tonne of popups after a few hours ;)

I don't think this needs to be separately configurable for Home and Root, since the default threshold is like 200 MiB (it's not a configured
percentage), so the absolute free space it warns about will be the same, even if your Home is giant compared to Root.

BUG: 340582
FIXED-IN: 5.20.0

Differential Revision: https://phabricator.kde.org/D29770