Use more compact OSD
Summary:
A frequent complaint over the years is the size of the OSD. It was tried to alleviate that by having it start fading out slowly immediately but the way it was done wasn't ideal, didn't work on Wayland, and also causes flickering issues in recent Qt versions.
This changes the OSD to a bar-like design similar to the one used in Plasma 4.

BUG: 344393
BUG: 372665
FIXED-IN: 5.20.0

Depends on D29263

Test Plan:
Various OSD messages
{F6773939}
It can grow, if necccessary, to accomodate translations, up to half the screen width.
With Air theme
{F6773940}
Full desktop screenshot for some context
{F6773941}

Reviewers: #plasma, #vdg, broulik, niccolove, ndavis

Reviewed By: #vdg, niccolove, ndavis

Subscribers: ndavis, kori, Armstrong, alexde, achauvel, abetts, ngraham, davidedmundson, hein, Codezela, filipf, zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20569