[Notifications] Position popups more relative to notification icon
The available screen rect for notification popups is horizontally restricted by the panel they're in.
This way when you have a centered panel, the notification popup are "emitted from the panel" rather
than floating in mid air. The horizontal edge (left or right) is determined by which half of the
panel the notification plasmoid is in. When in System Tray, the location of System Tray is relevant.

For a vertical panel, it changes the position from bottom to top based on which half of the screen
the notification popup is on.

This affects only the default "near notification icon" placement setting. Manually setting a screen
corner still moves the notification popup all the way to the screen corner.