[Notifications] For group parents combine the child bodies
This enables some basic grouping when pointing a list at the group parents instead of the individual child elements.
Effectively unused right now (popups don't group and history only shows the application name for sections)
but this is in preparation for being able to have groups in the popup as well.

Differential Revision: https://phabricator.kde.org/D27130