BookmarksRunner: Change caching databases
Summary:
The copying/updating of the cache has been removed from the FetchSqlite class.
This allows the browsers to have handle the caching differently.

The firefox sqlite files are only copied if they changed. Before they were copied
for each match session.

The chrome favicon database is also only copied if needed
and the check is also made in the prepare method and not just when
the profile is initialized.

Test Plan:
For both firefox and chrome:
Plugin shows icons for results.
Add new bookmarks, close browser, the new bookmarks should show up with icons.

Reviewers: #plasma, ngraham, meven, broulik

Reviewed By: meven

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29726