[applets/mediacontroller] Don't use Album Art for icon
Summary:
Icon of Media Player applet is missing in the System Tray settings when
the media playing has a cover. Such icon is present when nothing is playing.
Fix by not using album art in any case - icon is not longer used in
tooltip, can be static.

BUG: 420197

FIXED-IN: 5.19.0

Test plan:
1. play audio/video with a player compatible with Media Player applet
2. open system tray settings during playback
3. Icon of Media Player applet is always present in system tray
settings, regardless of state.