Bookmarks Runner: Fix extraction of firefox profile
Summary:
BUG: 418526

The profile group with the value `Default=1` is not always the default profile, for example in the profiles.ini file from the bug report:
```
[Profile1]
Name=default
IsRelative=1
Path=dun5pnnn.default
Default=1

[Profile0]
Name=default-default
IsRelative=1
Path=xewbk2mp.default-default

[Install11457493C5A56847]
Default=xewbk2mp.default-default
Locked=1```

In the text you can see that the profile with `dun5pnnn.default` has the value `Default=1`,
but it is not  the default profile.
The correct default profile is in the `[Install11457493C5A56847]` group with the key `Default`.

Test Plan:
If you create a new profile it will be extracted as the default profile
(but the old value might still be cached see D28619).

Reviewers: #plasma, ngraham, bruns, broulik, meven

Reviewed By: meven

Subscribers: meven, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29326