Fix calendar events not being shown at first
I noticed that on the first time I bring up the calendar by clicking the clock on my Plasma panel, it always shows 'No event for today'. If I click another day on the calendar and then click back on today, it shows today's events fine. It should display it correctly the first time. To fix this bug, we need to compare the date without time. This patch is the same as the one that fixes BUG 395956, but only for Plasma/5.18 branch.