[applets/systemtray] Fix context menu misplaced
For SNI icons menu is imported from DBus. If during import it is
changed, for example some actions added or replaced, new actions are
added, but old one are only marked for removal in next event loop run.
As a result menu size is too big until delereLater takes effect.
Queue connection, so that Qt will have a chance to delete menu entires
before menu is used.

BUG: 421275
FIXED-IN: 5.18.6