Fix DBus service used in shutdown interface
During Plasma 5.18, ksmserver was cleverly split so it hosted the new
new interface registering both services names org.kde.ksmserver and
org.kde.Shutdown.

This way we could do a gradual port, update the libs, and migrate the
final code without any breakages. It was a good plan in theory.

In 5.19 we did the final actual splitting, unfortunately in
libkworkspace on the path where we skip logout confirmation had the
wrong name.

BUG: 423391