[applets/systemtray] Add a button in the hidden view to show the config window
Without this, the only way to show the config window for the System Tray
itself is by right-clicking, and as we all know, this is a no-no because
many users don't right-click on stuff and you can't right-click at all on
touch and we want our stuff to work on convertible devices with
touchscreens. So let's add a big beautiful button that people can click
on and touch.