Only spawn dbus-run-session if there isn't a session already
Typically on modern systems this will be started for us earlier in the
startup series and we don't have to worry about it.

For X11 clients look for the bus address on the X11 root window as a way
of sharing.

For reasons at the time we start our own DBus session within our wayland
session. Nesting causes conflicts and problems so this patch only start
a dbus session if one has not been set up already.

BUG: 404335