[OSD] Colorize percentage text when over 100%
We've now made it easier to increase the maximum volume to 150% for
those who need it due to quiet speakers. However since this could
potentially lead to excessive volume or damage to loud speakers with
long-term use, let's colorize the text to subtly indicate that this is
not a normal state.