[applets/batterymonitor] Handle "plugged in but still discharging" state
Right now we handle the "plugged in but not charging" state but not the
state where you're plugged in but the power supply doesn't supply enough
power to prevent discharging anyway. We can wiggle the logic around to
catch this case with the existing one for the "plugged in and not
charging" state.