set formfactors for applets that shouldn't be on phone
Many applets shouldn't actually appear on the phone, like
devicenotifier, the battery applet, the systray and so on
Populate all desktop files of the plasmoids with the proper
formfactors. on the desktop we keep showing everything.
on the phone most of those applets will be filtered out

Also take into account of the formfactors when populating the available
categories, wo we don't have the risk of seeing empty categories