[applets/systemtray] Return to pre-5.20 default panel icon arrangement
After much much much VDG discussion, it turned out that we don't
actually want to make the System Tray icons scale with the panel. We
returned to that with 31c7ebe319229e6834397c57c48e34a333bf4b88, but one
element remained: to make the icons automatically become a
two/row/column layout for thick panels, as we had for Plasma 5.19 and
earlier. This commit does that, and in the process simplifies the
configuration UI to show only the two options that we decided make sense:

1. Small icons that become two-row/column for thick panels
2. Auto-sizing icons that match the size of other panel widgets

Number one returns to being the default as it was before, and number two
is available for people/themes that want all icons to match in size.