[applets/devicenotifier] Remove UI to suppress popup  for new device connection
This option is just too nerdy to show a UI for. We keep the logic for it
in the code (it's just two lines) but it's now only available for people
who know about this and manually add "popupOnNewDevice=false" to their
plasma-org.kde.plasma.desktop-appletsrc file.

This prevents the applet from being too cluttered with options with the
new design.