[logout screen] Set implicit widths on OK and Cancel buttons
This was previously done in the component itself, but that was removed
in 711789908b0c6c0f76597596d567fba4ae9151ca (plasma-framework). For the
most part, buttons look better without the old minimum width, but these
buttons here relied on that behavior and do in fact look better with it.
So let's set it manually on these buttons.