Allow comma when dot is decimal separator
BUG: 406388
FIXED-IN: 5.20

This allows us to use a comma as a decimal separator, even if we have a
dot configured.
Especially when there is a difference between the official and
implemented decimal separator (see bug report) or if the user switches
between settings with a different separator.
By making sure that no [ or ] character(used for vectors) is contained
before replacing the string we make sure that we still provide all the
features Qalculate does.