Move "Configure Desktop" menu item to the top of the menu
This came up in today's VDG meeting. A user who right-clicks on the
desktop is highly likely to want to change the wallpaper, but the menu
item that does this is the last item in the menu, not the first. We can
move it up to the top.

The "Open with Dolphin" item is likewise moved up too so that we don't
have two single-item sections in a row.