krunner: Fix performance issue raised by PC3
Simplifies the logic of the busy indicator since now it will already
fade out when it's not running.
Also addresses the logic of the timer which was off.

This made krunner freak out on Wayland for some reason to investigate,
both opacity animations were triggering a repaint and we tried to render
too often making the process block. With this change it works fine
again.