[applets/batterymonitor] Add display of screen brightness as a percentage
At the moment, in the battery widget, the slider displays the brightness of the screen, but
it doesn't have text labels, so it's quite difficult to determine the brightness of the
screen. Adding a label over the slider that will display the current screen brightness as a
percentage will make this applet more descriptive and handy.