Add margins to System Activity window
Right now it is deliberately setting the margins to zero, which causes
all of its UI controls and framed areas to touch the window border,
which looks bad and is just an incorrect thing to do in general for a
thing that does not deliberately provide its own margins for content.

Let's not do that anymore, so the window looks better.

BUG: 426661
FIXED-IN: 5.20


(cherry picked from commit 42e2078cfd41396ceb6d78746f2ebbdd53b11ee8)