Work around missing geometryChanged update signals
In Qt5.15.1 geometryChanged signals are not emitted if
XCB_RANDR_SCREEN_CHANGE_NOTIFY and XCB_RANDR_NOTIFY arrive together.

The geometry property is still updated

See bug QtBug - 86604

This disproportionately affects plasma as we are quite screen-dependent
and are active when screens are doing the initial startup where both of
these properties change. So I think a temporary workaround is worth
considering.

This hack emits the missing signals from what is emitted. It will result
in a bit more noise, but most implementations should just no-op.

BUG: 426496