[applets/systemtray] Only show back button when the hidden items view has any items
Today on master, the system tray menu *always* shows a back button, regardless of the context.

But when all items are visible, the back button simply acts as a "close" button, since we are
only "one layer" deep and there is nowhere to go back to. This commit fixes that issue by only
displaying the back button hen the hidden items view has any items in it such that you could go
back to it.