[applets/devicenotifier] Don't open popup for already-connected devices
The old applet only opened the popup when a new device was attached
after the applet was loaded, not if there was already a device attached
when the applet was loaded. This commit brings back the old behavior by
re-using the approach taken in the old applet.

BUG: 426990
FIXED-IN: 5.20


(cherry picked from commit 7f5339cb77c77caba94088fb1c9e28a913c21eda)