kio_desktop: Fix creating a new folder while at "desktop:"
If you navigate to "desktop:" (without a / at the end) then create a new
folder, you'll end up with "/home/user/DesktopNew Folder", that's because
the path part of the url is empty. Change the code to always add / to the
beginning of the path when needed.

Drive-by change: silence a compiler warning about an unused parameter.