[libtaskmanager] Remove a poorly conceived optimization
The LauncherTasksModel tried to avoid a model reset by making certain
assumptions to save a few cycles. Ultimately this was misguided, in-
correctly not evicting the internal app data cache as well as causing
a lot more work in the view trying to compute layout deltas and some-
times causing a few frames of ugly animations.

There likely _is_ some optimization potential here, but it needs to
be done smarter later on.