[applets/systemtray] Switch to multi-row/column view at higher panel thickness
Right now he system tray displays a two row/column view at 50px, which
makes the two rows or columns feel a bit too squeezed.

This MR increases the threshold by adding some spacing, which changes
the cut-off to a 56px panel. As a result, it doesn't look as squeezed.