[Notifications] Also check transient parent for whether it's a dialog
In case a context menu is opened, we want to check the dialog it came from.
This fixes the case of notifications dancing around when opening a context menu in
System Tray but does not fix the case when opening a context menu in the notification
itself since there it cannot know that the still open System Tray popup is the one
we should evade and not the popup with the menu in it.

CCBUG: 426187


(cherry picked from commit bc12c6ca60331939c6070c93c63022368241d7c4)