[applets/devicenotifier] Don't offer to unmount non-removable devices
The previous check for whether a device was non-removable in the
delegate was faulty and was doing too much work, which caused it to miss
non-root internal disks. In fact, we can already get this data straight
from the data engine. This fixes the bug and removes code in net.

BUG: 427176
FIXED-IN: 5.20.1