[applets/icon] Handle non-Applications URLs again
In the past, this applet naively passed the URL to KRun when clicked on
or something was dragged on top of it. However during porting away from
KRun, it was ported to KIO::ApplicationLauncherJob, which only handles
applications, not other types of URLs.

This commit fixes that by making it use KIO::OpenURLJob (which can
handle applications as well as arbitrary URLs) when it's clicked on. We
also short-circuit the application-specific logic when something is
dragged on it and it's not an application, which allows the
non-application logic to be reached.

BUG: 427797
FIXED-IN: 5.20.3


(cherry picked from commit 87b562063b9580c3996c88654368f924a59a061a)