[Notifications] Add "Move to Trash" and "Delete" actions to file menu
So one can, for instance, quickly delete an unwanted screenshot.
It follows the usual Trash context menu actions, i.e. move to trash is only
shown for local files and locations that support moving, Delete is shown for
remote locations or when enabled in settings.

Swapping of the action when holding Shift is not implemented.

BUG: 428159
FIXED-IN: 5.21.0