[Notifications] Take application name from desktop name
In case no application name is set but we found a service, use the service name.
Also prefer the service name when the application name just looks like the desktop
entry, e.g. VLC just sets "vlc" as application name, which we would then override
to "VLC media player" from the vlc.desktop we might have found.

This is actually mostly just preparation for should we add a look up of the desktop
entry through the process CGroup.