Fix SystemEntries not updating correctly
The code is meant to add the entry if it's valid, then watch for
changes. Somehow we end up only monitoring if it was valid initially.
This doesn't make sense.

This means if the menu loads before the backend we don't update
correctly when it does load.

CCBUG: 423526
BUG: 427779


(cherry picked from commit 7c5f16c40e9bf27d6b1466c220b9eeaf800d4620)