[libkworkspace] Fix if getCurrentSeat needs to fallback to old approach
The test for whether "/auto" is supported was bogus.

QDBusAbstractInterface::isValid effectively just confirms the service
exists and the path is a legally valid name, not that it has any
interfaces there.

This means the fallback path is not run appropriately.

BUG: 423526


(cherry picked from commit f6269cadde64ac535092a74cb4fc493cc25014fe)