[Notifications] Check pause button when job is paused
To provide some more direct feedback that the job is paused and an easy to spot
place to click to resume.

I chose to check the button instead of changing the icon like is done in Media
Controller for more visual feedback. A job is also seldom paused so having it
more obvious when it is the case aids the user better.


(cherry picked from commit de6ac8e8fa7dd1ef2200776cc4bd1382a4e4ffad)